Move the FakeSummaryCache out of virtual/common into its own package.
